+++ modules/ballerina-core/src/test/java/org/ballerinalang/model/expressions/ArrayAccessExprTest.java
@@ -51,8 +51,8 @@ public class ArrayAccessExprTest {
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
 
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 210;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 210;
         Assert.assertEquals(actual, expected);
     }
 
@@ -84,8 +84,8 @@ public class ArrayAccessExprTest {
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
 
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 11;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 11;
         Assert.assertEquals(actual, expected);
     }
