+++ modules/ballerina-core/src/test/java/org/ballerinalang/model/expressions/MapAccessExprTest.java
@@ -53,8 +53,8 @@ public class MapAccessExprTest {
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
 
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 105;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 105;
         Assert.assertEquals(actual, expected);
     }
