+++ modules/ballerina-core/src/test/java/org/ballerinalang/model/expressions/NotEqualExprTest.java
@@ -49,8 +49,8 @@ public class NotEqualExprTest {
         BValue[] returns = BLangFunctions.invoke(bLangProgram, "checkIntEquality", args);
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 1;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 1;
         Assert.assertEquals(actual, expected);
 
         args = new BValue[] { new BInteger(5), new BInteger(6) };
@@ -70,8 +70,8 @@ public class NotEqualExprTest {
         BValue[] returns = BLangFunctions.invoke(bLangProgram, "checkFloatEquality", args);
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 1;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 1;
         Assert.assertEquals(actual, expected);
 
         args = new BValue[] { new BFloat(8.0001f), new BFloat(6.9992f) };
@@ -91,8 +91,8 @@ public class NotEqualExprTest {
         BValue[] returns = BLangFunctions.invoke(bLangProgram, "checkBooleanEquality", args);
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 1;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 1;
         Assert.assertEquals(actual, expected);
 
         args = new BValue[] { new BBoolean(true), new BBoolean(false) };
@@ -112,8 +112,8 @@ public class NotEqualExprTest {
         BValue[] returns = BLangFunctions.invoke(bLangProgram, "checkStringEquality", args);
         Assert.assertEquals(returns.length, 1);
         Assert.assertSame(returns[0].getClass(), BInteger.class);
-        int actual = ((BInteger) returns[0]).intValue();
-        int expected = 1;
+        long actual = ((BInteger) returns[0]).intValue();
+        long expected = 1;
         Assert.assertEquals(actual, expected);
 
         args = new BValue[] { new BString("This is Ballerina !!!"),
