+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/datatables/GetByIndexStringReturn.java
@@ -61,7 +61,8 @@ public class GetByIndexStringReturn extends AbstractNativeFunction {
 
     public BValue[] execute(Context ctx) {
         BDataTable dataTable = (BDataTable) getArgument(ctx, 0);
-        int index = ((BInteger) getArgument(ctx, 1)).intValue();
+        //Here we cast index value to int for consistency
+        int index = (int) ((BInteger) getArgument(ctx, 1)).intValue();
         String type = (getArgument(ctx, 2)).stringValue();
         return getBValues(dataTable.get(index, type.toLowerCase(Locale.ENGLISH)));
     }
