+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/system/LogInt.java
@@ -63,7 +63,8 @@ public class LogInt extends AbstractNativeFunction {
     private static final Logger logger = LoggerFactory.getLogger(LogInt.class);
 
     public BValue[] execute(Context ctx) {
-        LogUtil.log(logger, ((BInteger) getArgument(ctx, 0)).intValue(), getArgument(ctx, 1).stringValue());
+        //here we cast second parameter to int as anyway it only has few log levels
+        LogUtil.log(logger, (int) ((BInteger) getArgument(ctx, 0)).intValue(), getArgument(ctx, 1).stringValue());
         return VOID_RETURN;
     }
 }
