+++ modules/ballerina-native/src/test/java/org/ballerinalang/nativeimpl/functions/impl/TestConstantsNativeFunction.java
@@ -61,7 +61,8 @@ public class TestConstantsNativeFunction extends AbstractNativeFunction {
 
     public BValue[] execute(Context ctx) {
         // TODO : Improve this with trace log.
-        LogUtil.log(logger, ((BInteger) getArgument(ctx, 0)).intValue(), getArgument(ctx, 1).stringValue());
+        //here we cast second parameter to int as anyway it only has few log levels
+        LogUtil.log(logger, (int) ((BInteger) getArgument(ctx, 0)).intValue(), getArgument(ctx, 1).stringValue());
         return VOID_RETURN;
     }
 }
